+++ include/extras-kwidgetsaddons/libkled.h
@@ -1763,7 +1763,7 @@ bool k_led_is_full_screen(void* self);
 /// @param self KLed*
 ///
 /// @return flag of enum Qt__WindowState
-int64_t k_led_window_state(void* self);
+int32_t k_led_window_state(void* self);
 
 /// Inherited from QWidget
 ///
@@ -1771,7 +1771,7 @@ int64_t k_led_window_state(void* self);
 ///
 /// @param self KLed*
 /// @param state flag of enum Qt__WindowState
-void k_led_set_window_state(void* self, int64_t state);
+void k_led_set_window_state(void* self, int32_t state);
 
 /// Inherited from QWidget
 ///
@@ -1779,7 +1779,7 @@ void k_led_set_window_state(void* self, int64_t state);
 ///
 /// @param self KLed*
 /// @param state flag of enum Qt__WindowState
-void k_led_override_window_state(void* self, int64_t state);
+void k_led_override_window_state(void* self, int32_t state);
 
 /// Inherited from QWidget
 ///
@@ -2296,7 +2296,7 @@ void k_led_render3(void* self, void* target, void* targetOffset, void* sourceReg
 /// @param targetOffset QPoint*
 /// @param sourceRegion QRegion*
 /// @param renderFlags flag of enum QWidget__RenderFlag
-void k_led_render4(void* self, void* target, void* targetOffset, void* sourceRegion, int64_t renderFlags);
+void k_led_render4(void* self, void* target, void* targetOffset, void* sourceRegion, int32_t renderFlags);
 
 /// Inherited from QWidget
 ///
@@ -2326,7 +2326,7 @@ void k_led_render32(void* self, void* painter, void* targetOffset, void* sourceR
 /// @param targetOffset QPoint*
 /// @param sourceRegion QRegion*
 /// @param renderFlags flag of enum QWidget__RenderFlag
-void k_led_render42(void* self, void* painter, void* targetOffset, void* sourceRegion, int64_t renderFlags);
+void k_led_render42(void* self, void* painter, void* targetOffset, void* sourceRegion, int32_t renderFlags);
 
 /// Inherited from QWidget
 ///
@@ -2343,7 +2343,7 @@ QPixmap* k_led_grab1(void* self, void* rectangle);
 /// @param self KLed*
 /// @param typeVal enum Qt__GestureType
 /// @param flags flag of enum Qt__GestureFlag
-void k_led_grab_gesture2(void* self, int64_t typeVal, int64_t flags);
+void k_led_grab_gesture2(void* self, int64_t typeVal, int32_t flags);
 
 /// Inherited from QWidget
 ///
